change base dtype from int8 to int16 to avoid overflow when cube size is >= 7

+++ src/rubik/action.py
@@ -15,7 +15,7 @@ POS_ROTATIONS = torch.stack(
                 [0, 0, 0, 1],
                 [0, 0, -1, 0],
             ],
-            dtype=torch.int8,
+            dtype=torch.int16,
         ),
         # rot about Y: X -> Z
         torch.tensor(
@@ -25,7 +25,7 @@ POS_ROTATIONS = torch.stack(
                 [0, 0, 1, 0],
                 [0, 1, 0, 0],
             ],
-            dtype=torch.int8,
+            dtype=torch.int16,
         ),
         # rot about Z: Y -> X
         torch.tensor(
@@ -35,7 +35,7 @@ POS_ROTATIONS = torch.stack(
                 [0, -1, 0, 0],
                 [0, 0, 0, 1],
             ],
-            dtype=torch.int8,
+            dtype=torch.int16,
         ),
     ]
 )
@@ -46,7 +46,7 @@ POS_SHIFTS = torch.tensor(
         [0, 1, 0, 0],
         [0, 0, 1, 0],
     ],
-    dtype=torch.int8,
+    dtype=torch.int16,
 )
 
 
@@ -74,7 +74,7 @@ def build_actions_tensor(size: int) -> torch.Tensor:
             for inverse in range(2)
         ],
         dim=0,
-    ).sum(dim=0, dtype=torch.int8)
+    ).sum(dim=0, dtype=torch.int16)
 
 
 def build_action_tensor(size: int, axis: int, slice: int, inverse: int) -> torch.Tensor:
@@ -87,7 +87,7 @@ def build_action_tensor(size: int, axis: int, slice: int, inverse: int) -> torch
     length = 6 * (size**2)
 
     # extract faces impacted by the move
-    indices = tensor.indices().to(dtype=torch.int8)  # size = (4, length)
+    indices = tensor.indices().to(dtype=torch.int16)  # size = (4, length)
     changes = (indices[axis + 1] == slice).nonzero().reshape(-1)  # size = (n,), n < length
     extract = indices[:, changes]  # size = (4, n)
 
@@ -97,7 +97,7 @@ def build_action_tensor(size: int, axis: int, slice: int, inverse: int) -> torch
     rotated = rotated + offsets  # size = (4, n)
 
     # apply face rotation
-    rotated[0] = (F.one_hot(rotated[0].long(), num_classes=6).to(torch.int8) @ FACE_ROTATIONS[axis]).argmax(dim=-1)
+    rotated[0] = (F.one_hot(rotated[0].long(), num_classes=6).to(torch.int16) @ FACE_ROTATIONS[axis]).argmax(dim=-1)
 
     # from this point on, convert rotation into a position-based permutation of colors
     (inputs, outputs) = (rotated, extract) if bool(inverse) else (extract, rotated)
@@ -116,11 +116,11 @@ def build_action_tensor(size: int, axis: int, slice: int, inverse: int) -> torch
     # convert permutation dict into sparse tensor
     perm_indices = torch.tensor(
         [[axis] * length, [slice] * length, [inverse] * length, list(total_perm.keys()), list(total_perm.values())],
-        dtype=torch.int8,
+        dtype=torch.int16,
     )
-    perm_values = torch.tensor([1] * length, dtype=torch.int8)
+    perm_values = torch.tensor([1] * length, dtype=torch.int16)
     perm_size = (3, size, 2, length, length)
-    return torch.sparse_coo_tensor(indices=perm_indices, values=perm_values, size=perm_size, dtype=torch.int8)
+    return torch.sparse_coo_tensor(indices=perm_indices, values=perm_values, size=perm_size, dtype=torch.int16)
 
 
 def parse_action_str(move: str) -> tuple[int, ...]:
